Move Around Stats and Add Health Component
Preparation for making attack damage work again.
