+++ src/fighter.rs
@@ -1,10 +1,12 @@
 use bevy::prelude::*;
 use bevy_rapier2d::prelude::CollisionGroups;
 use rand::prelude::SliceRandom;
+use serde::Deserialize;
 
 use crate::{
     animation::Animation,
     collisions::BodyLayers,
+    damage::{Damageable, Health},
     enemy::Enemy,
     fighter_state::{Idling, StateTransitionIntents},
     metadata::FighterMeta,
@@ -23,10 +25,31 @@ pub struct ActiveFighterBundle {
     character_bundle: CharacterBundle,
     #[bundle]
     physics_bundle: PhysicsBundle,
+    health: Health,
+    damageable: Damageable,
     state_transition_intents: StateTransitionIntents,
     /// Fighters start off idling, but this component may be removed when the fighter state changes.
     idling: Idling,
     velocity: LinearVelocity,
+}
+
+#[cfg_attr(feature = "debug", derive(bevy_inspector_egui::Inspectable))]
+#[derive(Component, Deserialize, Clone, Debug)]
+#[serde(deny_unknown_fields)]
+pub struct Stats {
+    pub max_health: i32,
+    pub damage: i32,
+    pub movement_speed: f32,
+}
+
+impl Default for Stats {
+    fn default() -> Self {
+        Stats {
+            max_health: 100,
+            damage: 35,
+            movement_speed: 17000.,
+        }
+    }
 }
 
 /// Turns a fighter stub data (loaded from the metadata) into a fully active fighter.
@@ -70,6 +93,8 @@ impl ActiveFighterBundle {
                 stats: fighter.stats.clone(),
                 ..default()
             },
+            health: Health(fighter.stats.max_health),
+            damageable: default(),
             physics_bundle: PhysicsBundle {
                 collision_groups: CollisionGroups::new(body_layers, BodyLayers::ALL),
                 ..default()
